Migrate to typescript + Bump minimatch + Add ESM support
