+++ .github/workflows/test.yml
@@ -17,6 +17,7 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
       - run: npm ci
+      - run: npm run build
       - run: npm test
   coverage:
     runs-on: ubuntu-latest
@@ -26,6 +27,7 @@ jobs:
       with:
         node-version: 20.x
     - run: npm ci
+    - run: npm run build
     - run: npm test
     - uses: coverallsapp/github-action@master
       with:
