+++ src/index.cjs.ts
@@ -1,0 +1,3 @@
+import {readdirGlob} from './index';
+
+export = readdirGlob;
