+++ tsconfig.cjs.json
@@ -1,0 +1,11 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "declarationDir": "dist/cjs",
+    "outDir": "dist/cjs"
+  },
+  "files": ["src/index.cjs.ts"],
+  "exclude": ["node_modules", "test/**"]
+}
