+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "lib": ["es2018"],
+    "target": "es2018",
+    "strict": true,
+    "moduleResolution": "node10",
+    "declaration": true,
+    "noImplicitAny": true,
+    "sourceMap": true,
+  }
+}
